requirements.txt
Flask>=2.2.0
psycopg2-binary==2.9.10
selenium==3.141.0
python-dotenv==0.19.1
chromedriver-autoinstaller==0.2.2
pandas==2.2.3
numpy==1.26.4
requests==2.31.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
resend==0.8.0